requirements.txt
anthropic
ollama
python-dotenv
openai
